refactor(execution): move linear memory into store submodule
