+++ src/execution/store/linear_memory.rs
